JBPM-4400 - resolve split package problem for human task audit
