+++ jbpm-human-task/jbpm-human-task-audit/src/main/java/org/jbpm/services/task/audit/impl/model/BAMTaskSummaryImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jbpm.services.task.impl.model;
+package org.jbpm.services.task.audit.impl.model;
 
 import java.io.Serializable;
 import java.util.Date;
