+++ jbpm-human-task/jbpm-human-task-audit/src/test/java/org/jbpm/services/task/audit/test/LifeCycleBaseTest.java
@@ -27,8 +27,8 @@ import org.jbpm.services.task.audit.DeleteAuditEventsCommand;
 import org.jbpm.services.task.audit.GetAuditEventsCommand;
 import org.jbpm.services.task.audit.commands.DeleteBAMTaskSummariesCommand;
 import org.jbpm.services.task.audit.commands.GetBAMTaskSummariesCommand;
+import org.jbpm.services.task.audit.impl.model.BAMTaskSummaryImpl;
 import org.jbpm.services.task.impl.factories.TaskFactory;
-import org.jbpm.services.task.impl.model.BAMTaskSummaryImpl;
 import org.junit.Test;
 import org.kie.api.task.model.Status;
 import org.kie.api.task.model.Task;
