+++ jbpm-human-task/jbpm-human-task-audit/src/test/java/org/jbpm/services/task/audit/test/LocalLifeCycleTest.java
@@ -20,7 +20,7 @@ import javax.persistence.Persistence;
 
 import org.jbpm.services.task.HumanTaskServiceFactory;
 import org.jbpm.services.task.audit.JPATaskLifeCycleEventListener;
-import org.jbpm.services.task.lifecycle.listeners.BAMTaskEventListener;
+import org.jbpm.services.task.audit.lifecycle.listeners.BAMTaskEventListener;
 import org.junit.After;
 import org.junit.Before;
 import org.kie.internal.task.api.InternalTaskService;
