+++ jbpm-human-task/jbpm-human-task-audit/src/test/resources/META-INF/persistence.xml
@@ -33,7 +33,7 @@
     <class>org.jbpm.services.task.impl.model.UserImpl</class>
     
     <!--BAM for task service -->
-    <class>org.jbpm.services.task.impl.model.BAMTaskSummaryImpl</class>
+    <class>org.jbpm.services.task.audit.impl.model.BAMTaskSummaryImpl</class>
     
     <!-- Event Classes -->
     <class>org.jbpm.services.task.audit.TaskEventImpl</class>
