+++ jbpm-human-task/jbpm-human-task-core/src/main/java/org/jbpm/services/task/persistence/TaskModelProviderImpl.java
@@ -1,0 +1,13 @@
+package org.jbpm.services.task.persistence;
+
+import org.kie.internal.task.api.TaskModelFactory;
+import org.kie.internal.task.api.TaskModelProviderService;
+
+public class TaskModelProviderImpl implements TaskModelProviderService {
+
+	@Override
+	public TaskModelFactory getTaskModelFactory() {
+		return new JPATaskModelFactory();
+	}
+
+}
