+++ jbpm-runtime-manager/src/test/resources/META-INF/persistence.xml
@@ -40,13 +40,14 @@
     <class>org.jbpm.services.task.impl.model.EmailNotificationHeaderImpl</class>
     <class>org.jbpm.services.task.impl.model.PeopleAssignmentsImpl</class>
     <class>org.jbpm.services.task.impl.model.ReassignmentImpl</class>
-    
+
     <class>org.jbpm.services.task.impl.model.TaskImpl</class>
     <class>org.jbpm.services.task.impl.model.TaskDataImpl</class>
     <class>org.jbpm.services.task.impl.model.UserImpl</class>
+
     
     <!--BAM for task service -->
-    <class>org.jbpm.services.task.impl.model.BAMTaskSummaryImpl</class>
+    <class>org.jbpm.services.task.audit.impl.model.BAMTaskSummaryImpl</class>
     
     <!-- Event Classes -->
     <class>org.jbpm.services.task.audit.TaskEventImpl</class>
