+++ jbpm-services/jbpm-executor/src/test/filtered-resources/META-INF/persistence.xml
@@ -71,7 +71,7 @@
     <class>org.jbpm.services.task.impl.model.UserImpl</class>
     
     <!--BAM for task service -->
-    <class>org.jbpm.services.task.impl.model.BAMTaskSummaryImpl</class>
+    <class>org.jbpm.services.task.audit.impl.model.BAMTaskSummaryImpl</class>
 
     <properties>
       <property name="hibernate.max_fetch_depth" value="3" />
