+++ jbpm-services/jbpm-executor/src/test/resources/META-INF/persistence.xml
@@ -76,7 +76,7 @@
     <class>org.jbpm.services.task.impl.model.UserImpl</class>
     
     <!--BAM for task service -->
-    <class>org.jbpm.services.task.impl.model.BAMTaskSummaryImpl</class>
+    <class>org.jbpm.services.task.audit.impl.model.BAMTaskSummaryImpl</class>
 
     <properties>
       <property name="hibernate.max_fetch_depth" value="3" />
