+++ jbpm-services/jbpm-kie-services/src/test/resources/META-INF/persistence.xml
@@ -29,13 +29,13 @@
     <class>org.jbpm.services.task.impl.model.EmailNotificationHeaderImpl</class>
     <class>org.jbpm.services.task.impl.model.PeopleAssignmentsImpl</class>
     <class>org.jbpm.services.task.impl.model.ReassignmentImpl</class>
-    
+
     <class>org.jbpm.services.task.impl.model.TaskImpl</class>
     <class>org.jbpm.services.task.impl.model.TaskDataImpl</class>
     <class>org.jbpm.services.task.impl.model.UserImpl</class>
     
     <!--BAM for task service -->
-    <class>org.jbpm.services.task.impl.model.BAMTaskSummaryImpl</class>
+    <class>org.jbpm.services.task.audit.impl.model.BAMTaskSummaryImpl</class>
 
     <class>org.drools.persistence.info.SessionInfo</class>
     <class>org.jbpm.persistence.processinstance.ProcessInstanceInfo</class>
