+++ jbpm-test/src/main/resources/META-INF/persistence.xml
@@ -44,7 +44,7 @@
     <class>org.jbpm.services.task.impl.model.EmailNotificationHeaderImpl</class>
     <class>org.jbpm.services.task.impl.model.PeopleAssignmentsImpl</class>
     <class>org.jbpm.services.task.impl.model.ReassignmentImpl</class>
-    
+
     <class>org.jbpm.services.task.impl.model.TaskImpl</class>
     <class>org.jbpm.services.task.impl.model.TaskDataImpl</class>
     <class>org.jbpm.services.task.impl.model.UserImpl</class>
@@ -52,7 +52,7 @@
     <class>org.jbpm.services.task.audit.TaskEventImpl</class>
     
     <!--BAM for task service -->
-    <class>org.jbpm.services.task.impl.model.BAMTaskSummaryImpl</class>
+    <class>org.jbpm.services.task.audit.impl.model.BAMTaskSummaryImpl</class>
     
     
     <properties>
