+++ jbpm-test/src/test/resources/META-INF/persistence.xml
@@ -47,7 +47,7 @@
         <class>org.jbpm.services.task.impl.model.UserImpl</class>
     
         <!--BAM for task service -->
-        <class>org.jbpm.services.task.impl.model.BAMTaskSummaryImpl</class>
+        <class>org.jbpm.services.task.audit.impl.model.BAMTaskSummaryImpl</class>
         
         <class>org.jbpm.services.task.audit.TaskEventImpl</class>
         
@@ -116,7 +116,7 @@
         <class>org.jbpm.services.task.impl.model.UserImpl</class>
     
         <!--BAM for task service -->
-        <class>org.jbpm.services.task.impl.model.BAMTaskSummaryImpl</class>
+        <class>org.jbpm.services.task.audit.impl.model.BAMTaskSummaryImpl</class>
     
         <exclude-unlisted-classes>true</exclude-unlisted-classes>
